feat: update Motoko recipe to use Mops (#5)
Updates the Motoko recipe to use Mops in place of dfx to manage the
compiler version and dependencies.

Also adds an `args` parameter to the recipe to pass additional compiler
flags, equivalent to dfx.

BREAKING CHANGE: renames `entry` to `main` and expects a `mops.toml`
file in the same directory as `icp.yaml`.